+++ requirements.txt
@@ -1,2 +1,9 @@
-pyyaml==5.3.1
+PyYAML==5.3.1
 requests==2.22.0
+google-api-core==2.12.0
+google-api-python-client==2.102.0
+google-auth==2.23.2
+google-auth-httplib2==0.1.1
+google-auth-oauthlib==1.1.0
+googleapis-common-protos==1.60.0
+
